Update metadata.yaml
Corrected  "List of exporters to be used for tracing, separated by commas. Default is otlp. none means no autoconfigured exporter." with "List of exporters to be used for logging, separated by commas. Default is otlp. none means no autoconfigured exporter." for the OTEL_LOGS_EXPORTER and the same description for OTEL_METRICS_EXPORTER with "List of exporters to be used for metrics..."